+++ packages/@codotype/ui/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Sized down `GenerateCodeButton`
 * Merged `ConfigurationEditor` component template and logic into `BlueprintEditor` component
 * Added `ProjectEditButton` component
+* Split `SchemaEditor` out of `BlueprintEditor` as a standalone component
 
 ---